backend/requirement.txt
fastapi[all]==0.110.1
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-jose[cryptography]==3.3.0
requests==2.32.3
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
